Added scripts and patches to append and update recoveries.
Added scripts that connect to the GLFC lamprey database and process
the lift(gear) and biosample tables into RecoveryEvent and Recovery
objects and insert them into the database.  This commit also includes
some patches and an associated python script that updates the
recoveries with additional information that not available in the
lamprey request (most notably, cwt sequence and micro-mark tags).  The
processing documentation has been updated to reflect the new steps.