Added model, factory and tests for Jurisdiction
Added a model to capture jursdictions by lake for each agency.
Technically, this could be handled as the middle table in a
many-to-many join between states/prov and lake.  The table also
includes a couple of geometry fields to capture polygond associated
with the juristiction shoreline/shape, and the extent of stocking
events.  The test verifies that the factor and str() method work as
expected.  All existing tests pass.