Added jurisdiction to stocking events and filters
Added jurisdition to the stockingevent model and removed direct
forgein keys to lake or state. These were replaced with property
methods on the stockingevent objects.  Added jursdiction as one of the
filters for stockingevent filter objects. Added elements to template
to render hyperlinks for juristiction as tags.  Added url to for
stocking events by jurisdiction and jursdiction by year.  Updates to
factories to match models.  All existing tests continue to pass.