Created ipython settings and fixed typo in reocvery model
The local settings file now includes settings for debug-toolbar that
causing errors when a local shell was opened.  the setting file
ipython can now be used for console interaction, local.py used for
running the development webserver.  Also fixed a typo in the recovery
model - not sure how that made it through without being caught.