Added piechartMap view, urls, and template
Added a PieChartMap as a class based view to serve as a container for
our piechartmap vizualization.  The the view collects the attributes
of the selected region, gets the selected year, finds the approiate
data url and loads the template.  This commit also includes an update
to production requirement files.