Rescale pie charts to a single maximum and added clearPointInfo()
The radius scale for our pie charts is now based on the grand total in
the presented view. Previously, they were scaled so that there was
always at least one circle at the maximum diameter, and all other
scaled accordingly. Now they are scaled as though there is a single
large pie chart (that may not even exists) - the result is, that the
number of pie chart pixels remains constant as the spatial resolution
is changed, providing an accurate indication of stocking over space.
Also added a method to clear the point info div - before it would
remain populated, event if the selected pie was long gone, at which
point it couldn't be easily turned off.