Updates to base template - Admin options and warning message
Updated the admin options and greeting message.  Admin options are only
available once the user is logged in. Options to change their
password, view the django admin, or logout are available. A Login link
is available if the user is not currently authenticated. A warning
message has been added to the base template, and currently appears on
every template, indicating that this web site is currently under
active development and is know to be incomplete at this time.