Added STATIC_ROOT and production.py
Preparing for deployment.  Added STATIC_ROOT setting so that we can
run collect static and get all of our assets in the right spot for
deployment.  Added basic production.py file.  This needs to be
carefully checked and verified, but contains most of the required
elements, and excludes those that are for local development
only (debug-toolbar).