Improvements to Check box Array
Small changes to the mark-up associated with each checkbox so that
they use semantic-ui formatting and css. Converted the selectAll and
Clear All links to buttons and moved them to the bottom of the
checkbox list, and pulled the select all button to the right. Changes
in the style sheet override semantic-ui default for accordions and
buttons to ensure that there is sufficient room for all elements.