Updated chechbox array to put buttons in their own div
The When we moved the check boxes to the bottom of the checkbox
arrays, they were still within the same dive as the checkboxes - when
checkboxes were added back onto the array (as other filters were
applied and then removed), the new check boxes were being added onto
the bottom of the div - below the buttons. This commit includes a new
template element for the buttons - it will alway be below all of the
checkboxes, no matter how many are added or removed.