Added functionality to change slice selector variable
Added the basic functionality to allow the user to dynamically change
the variable represented by the slices of the pie charts. The default
is species, but other categories such as agency, strain, life stage
and stocking methods are available too. Still need to update other
components of the user interface, but for the most part it works as expected.