Starting to make find events form dynamic
There are a large number of options that people can select from the
drop down select boxes and those selections effect the number of
available options in other boxes.  This is our first attempt at
dynamically updating those options (and estimating the number of
events returned) as the select widgets are updated. This commit also
includes an updated rollup.config.js file to watch and compile
multiple source and target files.