Dynamic dropdowns and crossfilter mostly working now.
This commit includes all of the code to load the unique stocking
events into a crossfilter and update the available choices in
dropdowns as other options are selected. the Total number of selected
events is also include on the page.  It still does not include any
form validation, or lead to anything when submitted.