Added generic API view to render found stocking events
Most of the changes in this commit are associated with efficiently
retrieving individual stocking events selected through the find
stocking events form. The events filter was updated to include in()
clause, and a utility function was created to convert cleaned form
data to correctly formatted url query parameters.  Some of the updates
are also associated with improvements to the find_events.js.