Added view for filtered events and setup associated js file
Added a separate view for filtered events that includes the query
parameters in url. The maximum number of records included in the
filtered results can be controlled from a global value in the django
settings file. If the number of records return by the ajax call is the
same as the maximum number, a warning message is displayed.