Updated filters div, added month plot and tweaked stacked bar plot
The filters div on the html page now reflects the charts and the
current filter statuses. A stocking month plot has been added to the
stocking event attributes row, and the x-axis of the stacked bar chart
was tweaked to ensure that the bars are not clipped and are labeled
only once regardless of how many years are selected.