Updated Species Pie Chart Colour Map and API Permissions
This commit includes updates to the main page pie chart color map so
that the species names are consistent with species names in the
database (which are now all title case). It also includes updates to
the permission classes of several of the api endpoints - the read only data needs be
available for both logged-in and anonymous users in order for the maps
and charts work as expected.