Added map+stats panel to template, generalized stacked bar plot
Added a map, stats panel, and radio button controls to the html
template for filtered stocking events.  Generalized reducer to accept
column variable, and updated the stacked bar graph to remove any
reference to species - what is plotted will be controlled by a category
variable that can be changed by radio buttons on the sidebar.