Stacked Barchart is now dynamic
The category that is represented by the stacked bar graph can now be
controlled by the radio buttons in the side bar. These buttons will
also control the category for slice of map pie charts when they are
eventually implemented.  The sidebar has been wrapped in a sticky so
that it remains in the view port when the user scrolls down the page.