Updated update_stat_panel props - what and label
The props passed into update stats panel uses to use something called
slices which was an array of objects with specific keys. A single
value, (label) was extracted from this array inside the stats summary
function. This extraction as been moved out of the function and
replaced with a 'label' argument.  The same function is used in the
filtered events code.