Strain uses slug as key and common color scale for selected variable
Strain has been modified to use a slug, which is a combination of the
strain code and species abbreviation instead of the strain id (which
doesn't mean anything to a human).  This commit also includes changes
that make the colour scale of the dc charts change if they are
selected.  This way, the pie charts, the stacked bar charts, the
legend, and the dc chart for the selected category are all the same.
The other dc charts use the default colour scheme.