Moved reducers from filteredevents.js to utils
Modified main.js and filteredevents.js to use the same reducers in
reducers.js. The new reducers are wrapped in a closure so that a
column argument can be passed to them, rather than relying on a global
variable. Had rename the "total_stocked" element in main.js to "total"
to ensure a consistent api.