Moved clips and marks to lookup and improved comments
Moved the concordance arrays for clips, marks and tags to a separate
js file. This will be replaced when we integrate these fields into the
database and event serializer.  Also moved species colour lookups to a
separate table so that the same colour scheme can used for all
graphics across the appliction.  Reviewed comments in
filteredStockngEvents.js and augmented them were possible.