Updated the xlsEvent form to use choices dictionary
Trying to improve performance - choices for the select widgets are
executed once in the view and passed to the form as a
dictionary (rather than once for each form in the formset).  The
dictionaries for grids and stat_districts are keyed by lake and then
filtered when the form is rendered.