Removed include from xlsEvents template & added loaded to file form
The original xlsEvents form template had an include for each cell of
the table, resulting in thousands of templates being rendered for even
small uploads.  This commit removes those include directives, but does
not appear to improve performance noticeably.  This commit also include
adds data-validate tags to each of the widgets to facilitate front end
validation.  This commit also includes a data-load on the file upload
form when the form is successfully submitted (prevents double submissions
and helps pass the time).