Basic xlsx data upload now works
This is a big commit that was a long time coming, and there are still
a lot of very rough edges, but we now have a basic upload
functionality working.  A perfectly clean spreadsheet can be loaded,
viewed in and html form and submitted.  There limited error checking,
and there is no way to view details of uploaded records yet, but the
basic functionality now works.