Changed upload event detail to use slug instead of id
modified the data upload url, view, and template to use the slug of
data upload events rather than the id to produce readable urls.  Also
updated the form to grab the lake and agency values from the first row
of the form to create the datauploadevent object. There are issues
with updates to the slug after the fact - they don't seem to be
reflected in the saved object.