Some tests for xls upload
Still a  lot to do, but this commit includes several tests associated
with processing the xls file upload.  Empty lists have been added to
the choice files just in case a key cannot be found and a custom html
select widget has been added to insert and disable values that are not
valid choices. Allows them to be passed in from xls file, where they
will be caught and flagged in the htlm template.  Most of the other
changes in this commit are formatting changes made by black.