Tests for validate_upload utility function
Created a parameterized test to verify that the validate_upload
utility function works as expected.  This test uses that same list of
invalid spreadsheet as integration test that verifies messages are
displayed when an invalid spreadsheet is uploaded.  This list of
invalid spreadsheets has been moved to a fixtures file so it can be
reused as needed.