Added slug to Strain model and test for common_lookups_api
Added a slug and custom save method to the Strain model to that
strains can be referenced by their species and strain code.  Added a
test to verify that the common_lookup_api returns an object with the
expected keys, and that each of the elements in those keys has the
expected structure.