Added interactive map for stocking event upload details
This commit adds an interactive map to the stocking uploads event,
displaying where each of the events occurred and what species they
were. There is also some interactively - stocking events in the table
can be identified by clicking on the points on the map - the selected
rows remain, while others are hidden. Click on the same point toggle
the selected mark, and all of the events re-appear. An api end point
was created to return the stocking events associated with a upload
event.  Most of the changes to js files were caused by prettier
automatically reformatting saved files.