Added labels and widths and map to edit stocking event form
This commit includes a number of small aesthetic tweaks to the
stocking event form. It now has more meaningful labels on each input,
and inputs are scaled to with more appropriate for their content.  The
form also includes a map that displays where the the stocking event is
currently plotted.