Basic integration tests for stocking event form and 404/500 pages
This commit includes some basic integration tests to verify that the
form to edit stocking events is only available to authenticated users,
that a 404 page is rendered if a matching stocking event cannot be
found, and that the main elements of the form appear on the page.  All
of the previously passing tests continue to pass (although there are
two that continue to fail). This commit also includes simple templates
for 404 and 500 errors.