js files for stocking_event_form
Moved js for map and future interactions for the stocking event form
to a separate js file and updated the roll-up config file
accordingly. The interactivity is currently limited to updating the
point if the map is clicked of the input boxes change, and updating the
input boxes if the map changes.