Created Spatial lookup API for lake, mu, jurisdiction and Grid10
This commit includes new views, utils and tests associated with a
spatial lookup, that will allow us to get the lake, management unit,
jurisdiction and/or 10-minute grid associated with a particular
point (lat-lon coordinate pair).  This commit also includes database
migrations that standardize the name of spatial fields to geom and
updates to associated scripts (geom.shoreline is now obj.geom).