Added description to html and made request data element array
Added a segment to the spatial lookup page that contains a description
of how to use the page and what it does.  Also modified the success
and error functions of request for lake, grid10 and jurisdiction so
that the data is returned as an array or an empty array, which is
handled properly by d3.