JS validation for date fields on stocking event form
This commit includes updates to get form validation working for date
fields on stocking event form.  The form now includes a calendar
widget that can be used to selected event date, or dates can be
manually entered using day, month, year fields. If the date is valid,
the calendar will be updated, if the entered date is not valid, errors
will be raised and the calendar cleared.