Created get_spatial_attr ape endpoint
Created an endpoint to return the basic information for the lake,
jurisdiction, stat_district, and grid10 for a provided lat-long.  The
api endpoint will be used by the stocking event detail form and event
upload form to verify that the spatial widgets are consistent with the
provided coordinates.  Tests were also created to verify that the api
endpoint returned the expected result.