Updates to get addError and removeError functions working
There were a number of challenges to getting the addError and
removeError functions to work properly and not conflict when several
errors were added and removed. This version seem to work although an
empty residual popup remains after all of the errors have been
fixed. I can live with that.