Renamed FinClip to CompositeFinClip and created m2m with event
Originally, fin clips were captured in a lookup table that contained
each of the composite fin clips in the spreadsheet.  This is an awkward
way to handle data that is really a many-2-many relationship.  This
commit includes changes to the model files, the associated migrations
and tests to rename fin clips to Composite clips, create a new fin
clips model and associated it with stocking events through a many to
many relationship.  Most of the tests currently pass, although all of
the associated with the form and event upload are failing.