Added some methods and tests to the StockingEvent model
This commit includes a number new or updated model methods or
properties to return clips, marks, and latlong_flag. Tests are
included to verify that the methods work as expected.  Most of the
existing tests continue to pass, although there are two failing
because they still need to be written.