Added get_composite_finclip method to stocking event model
This commit includes a new method on the stocking event model that
will get or create a related composite clip object. The table
composite clips is intended to reduce the number of database queries
and ensure consistency when multiple clips are applied.  An associated
tests ensures that the method works as expected, and returns NO and UN
as appropriate when those values are specified as clips.