Added tests to verify that cwt details are rendered
Added a couple of tests to verify that the cwt details are rendered
appropriately on the event detail page when they are associated with
an event, and not rendered when there are no cwts associated with the
event. All added a test to verify that the has_sequential_cwt property
of the stocking event model works as expected.  There are three
failing tests that are there as reminders for calculation of yearling
equivalents and latlon_flag.