get_or_create_cwt_series utility function and tests
This commit introduces a new utility function that will get or create
cwt_series (and associated cwt objects).  If existing objects are
available they will be used, otherwise they will be created. A test
file verifies that the function works as expected for new and existing
objects, as well as different tag types and manufacturers. This commit
also include changes to the CWT model - manufacturer is now included
in the unique constraint, slug and str() method. The slug field had to
be increased in size.  All of the previously passing tests continue to pass.