Added cwt_series formset, utilities, and tests to StockingEvent Form
This commit includes a formset for cwt series objects that is embedded
into the stocking event form.   The commit includes a number of
associated utilities to prepare cwt data or retrieve or create cwt
objects as needed. The commit also includes a number of tests that the
formset and utility functions work as expected.  All of the previously
passing tests continue to pass. Tests that were associated with cwt
field in the stocking event form have been removed.