Added cwt_formset to Stocking Event Form
This commit includes a formset that will create or delete cwts
or sequential associated with a stocking event.  The commit includes
utility to get or create sequential cwts, as well as tests to ensure
that everything works as expected.  cwt_numbers is not longer an
attribute of the stocking event form as it has been replaced by the
formset. The commit also include javascript in the stocking event form
to add additional forms to the formset as handle some validation logic
depending on cwt attributes.  Other js files were updated by
prettier.  All of the previously passing tests continue to pass.