Redirect user to homepage on login
This commit redirects users to the bookmark they have designated as
their homepage. If they don't have a homepage, they are re-directed to
the standard basin wide map with the most recent data.  This commit
includes two new tests to verify that users with and without bookmarks
are redirected properly.