Update crossfilter dimensions on piechart maps from url
This commit applies functionality to apply filters encoded in the url
to the crossfilter on page load.  The global filter object is also
updated accordingly so that the check-boxes are rendered correctly and
highlighted to indicate the presence of a filter.  Preliminary tests
indicated that bookmarks can be created and returned without any loss
of state.