Dynamic behaviour for previous and next year buttons
This commit adds JavaScript to the basin wide pie chart to maintained
applied filters from one year to the next.  Additionally, the
available years are included in the template and passed to the
JavaScript so that the Next Year and Previous Year buttons can be
enabled and disabled appropriate to keep users from navigating to a
year that does not have any data.  All of the previously passing tests
continue to pass.