Update basin wide map to reflect state encoded in url parameters
This commit includes updates to the main basin wide pie charts
JavaScript and the associated polygon overlays so that they are
initialized to reflect the state of variables when the page loads - the
correct filters are applied and the map is zoomed to the appropriate
spatial scale. Changes where made to the update_crossfilter function
to only change the spatial resolution if is larger than the current
scale (if we zoom to management unit, but have lake wide pie charts
that's not really meaningful.  but we can have 10-minute grid points
on a lake wide map).