Added Breadcrumb to page reload and manUnit to common lookup
This commit dynamically build breadcrumbs associated with mapstate
depending on feature/scale encoded in the url.  The common lookup
serializer (and associated tests) where updated to include management
unit so that we could traverse the lake-jurisdiction-management unit
hierarchy.  The associated test was updated to include an element for
management units. All of the previously passing tests continue to pass.