Added role, agency, lakes, and permissions to User objects
Attributes of role, agency, and lake(s) to the user objects and a
utility function has been created to allow different levels of CRUD
permissions.  A normal user cannot create, edit, or delete events,
stocking coordinators will be able to create, edit or delete their
own, and the great lakes stocking coordinator will be able to edit,
delete, or create objects for any lake or agency.  Several tests are
also included in this commit to ensure the utility function works as
expected. This commit also includes an update to the email field to
always save it to lowercase so that Foo@Bar.com is the same as
foo@bar.com.