Added integration tests for roles and permissions
This is a big commit that include more than 20 integration tests that
verify that elements associated with roles and permissions behave as
expected depending on the authorization of the user. It also includes
updates to the views and templates, as well as fixes to a number of
tests that were failing now that different roles exists. All of the
previously passing unit tests continue to pass.