Stock_id is now generated from creation year and pk
Stock_id is the human friendly id field that will be carried over from
the existing ms access database. It must be unique to a single
stocking event.  The existing convention has been to assign stock id
sequentially within a year and prefix the id field with the creation
year. This commit includes an update to emulate that pattern by using
the creation year and object pk. When an object is created, if a stock
Id exists, it will be used to maintain continuity with existing data,
if it is null, it will be temporarily replaced with a unique
string (uuid4) that is replaced the next time the object is
saved. Tests are included in this commit to ensure that this
behavior.  All previously passing tests continue to pass.