Added tests for region of interest api endpoint and event filter
This commit includes several tests that ensure the region of interest
enpoint returns the expected results if the roi is completely
disjoint, overlapping, or within other geometries.  It also include a
test of the roi filter verifying that it returns the expected stocking
events. All of the previously passing tests continue to pass.