Added xlsx download endpoint for stocking events
This commit includes the addition of an api endpoint which return a
spreadsheet (or json response) that closely matches the data
submission spreadsheet.  The commit includes the addition of the
drf-renderer-xlsx package, a new url, new serializer and view. It
also includes a test file to ensure that the endpoints return the
expected response and that the stocking event filters are working
properly.  All previously passing tests continue to pass.