Updates to stocking filter including template tags for event list
This is a big commits that refines the filters for stocking events,
updates the view and template, includes new template tags, and tests
to verify that all of the stocking event filters work as
expected. There are several failing, but all of the tests associated
with filtering stocking events pass.