Added NumberInOrNullFilter for stocking_month
A custom filter was added to accommodate values that are null so that
users can select stocking events (or cwts) that occurred in a
particular month OR no month.  99 is currently the placeholder for null
in filter, although this could be changed.  Tests where added for both
stocking_event_filter and cwtsequence_filters to ensure that the
months are correctly filtered including when month is null.