Finalized CWT_list and associated tests
This commit includes final touches to the cwt_list view to include
filter check boxes for each of the available filters including cwt
manufacturer and tag type.  It also includes some integration tests to
verify that the cwt_list render properly, with the expected elements
and respecting the applied filters.  The database setup and argument
list are shared with the cwt_filter tests and were moved up into
pytest_fixtures file.  The basic of cwt detail page have been created
so the urls on the cwt_list would work. The cwt details page is no
where near complete. All of the previously passing tests continue to pass.