Updates to scripts and utilities used to migrate from accdb
This commit includes changes the scripts and utilities that were used
to migrate the most recent version of the great lakes fish stocking ms
access file to fsdviz.  Most of these are throw-away code that won't
be needed once the final migration is complete. Committed here for
posterity.  All of the previously passing tests continue to pass.