Added test to verify elements of cwt details page render as expected
This commit includes several integration tests to verify that the
template associated with individual cwt numbers renders properly and
accounts for all of the combinations of warnings, sequential cwts, and
duplicate cwt numbers.  The new test pass. There are three tests
associated with the cwt formset that are now failing which seem to be
associated with how sequence ranges are specified.