Created api endpoint for cwt stocking events
This commit includes updates to api urls and views to create an
endpoint for cwt events.  The commit includes additional fixtures that
created re-used cwts - there continues to be bug in the cwt list that
confounds filters if cwts have been reused. The querysets need to be
generated from the stocking events - no cwt_sequences.  Updates to the
cwt filter has caused 4 previously passing tests to fail. They will be
addressed shortly.