Update the cwt-list view to use stocking event, not cwt sequence
The original version of the cwt stocking event list used cwtsequence as
the basis for all the filters and returned objects - this caused
issues with filtering where cwts have been re-used.  This commit
updates that view to use stocking events instead. This commit also
includes a parameterized test with  re-used cwts to ensure that the
filters are applied correctly and do not return events that do not
match the filter but are associate with a cwt that has been reused in
an event that does. All of the tests that are expected to pass, do
pass. Three continue to fail as reminders.