Added non-editable lower and upper fields to cwtsequence
Seq_lower and seq_upper value were added to the cwtsequence model so
that the range could be included in the serialized
responses (apparently numeric sequence is not serializable). The
are not editable, but are updated when the seq object is updated.
This commit includes updates to the model file, the associated
migration (which updates any existing ranges), and two tests that
verify the fields are updated when the object is saved.