Added non editable _dd_lat and _dd_lon to stocking event
This commit add two non-editable fields to the stocking event model
that are updated from the geometry when the stocking event is
saved. The geometry objects are problematic to serialize when used
outside of a model serializer (which can be slow). These fields can be
used instead and do not overwrite the reported lat-long. The commit
includes a migration to add the new fields and update them from geom,
as well as two unit tests that verify the fields are updated correctly
when stocking events are saved. All of the previously passing tests
continue to pass.