Updated StockingEvent Admin
Updated the stocking admin view so that it would be a little
performant and usable. The main stocking event form should still be
used for most editing purposes, but if something goes wrong, the
events will still be available in the admin. The order of the fields
was re-arrange the closely match the detail view, fields with a large
number of selects where updated to use autocomplete_fields, and
select_related was updated to minimize the number of associated
queries.