Added api endpoint for cwt xlsx download
Added a view, and simple serializer to download filtered cwt events to
xlsx. A button has been added to the cwt_list that will download of
the filtered cwts. Four additional tests were added to verify that the
downloaded xlsx response had the expected columns, and that any url
parameters (filters) are respected.  All of the previously passing
tests continue to pass.