Added files and updates associated with Find CWT view
This commit includes all of the updates and new files associated with
the view to find and visualize cwt stocking events. It includes
updates the urls and views, form, tests, and a js script that
visualizes the results.  It also includes a new validator for lists of
cwts and a small tweak to the settings file to reduce precision of the
leaflet form widget.  All of the previously passing tests continue to pass.