Changed _dd_lat and _dd_lon to geom_lat and geom_lon
The lat-long fields that were added to the stocking model to capture
the latitude and longitude from the geometry where poorly named and
causing issues in the templates. This commit renames those fields from
_dd_lat and _dd_lon to geom_lat and geom_lon. All of corresponding
changes where made in every associated file including the tests. All
of the previously passing tests continue to pass.