Added composite clip codes to common lookup api endpoint
Added the composite clip code to the api endpoint so that they could be
used in the data visualization templates. The composite clips lookup
is currently used only in the filtered cwt visualization but will be
need in the others shortly.  The api endpoint unit was updated to
include new element.