Update to all js assets include package.json and rollup.config
This commit is a large commit of all of the js assets, including the
files in the production build directory. All of the associated
dependencies in the package.json have been updated to current (or close
to current) versions.  The versions of the scripts in the static/ks
directory have been minimized for production.