Added basic instructions for xlsx data upload to template
This commit includes some basic instructions on the xls event upload
template to ensure users have required information - use the correct
template, run the validation first, ensure that there is only one lake
and agency in your submission, and that your submission includes fewer
than the maximum number of rows. A test was also included to ensure
that the instructions appear on the rendered page.  All previously
passing tests continue to pass.