Added id to state-province api endpoint
The validation mechanisms need a way to get state/province objects in
a way that is consistent with other spatial fields.  This commit add
id to the api response for state-province objects and updates the
tests accordingly. It also includes two new filters for management
units so that they can also be returned by the state_province
abbreviation or id.  Some tests associated with xlsx upload are
currently failing.