Added Role to Stocking Event Form
The widgets for agency and lake on the stocking event form need to
reflect the permission of the user - agency stocking coordinators
should only see lakes they are associated with and their agency. Great
Lakes Coordiantors should see all choices.  This commit includes those
updates to the form and updates to the tests to accommodate the new
kwarg argument and verify behavior depending on the user's role.