Updated the spatial validation of stocking event form
This commit updates the validation of the spatial input widgets -
jurisdiction now filters the list of available management
units. Additionally, the call to get the associated polygons based on
lat-lon works as expected and validates on page load.  All of the
previously passing tests continue to pass (although there are some
associated with the xlsx upload that are still failing).