Updated xlsx_form tests associated with lake
There were several tests that were failing because they were expecting
lake to be one of the available fields in the upload. Lake (and
agency) have been removed from the form because each upload bundle
must be associated with a single lake and agency.  All of the
currently available tests for the xlsx_form are now passing again.