Added validation to xls upload for dependent fields
This commit includes updates to the form template and associated js
scripts to update choices available in depend select
boxes (jurisdiction, stat_dist, grid, species and strain) and run
validation on blur event for each widgets.  There are currently
several tests failing that are associated with changes made to the
xls_form - strain is now a select widget, but it does not have any
available choices.  A good place to start on Monday....