Updates associated with Strain lookups for xls upload form
This commit includes updates to the common api lookup api and xls
upload validation to accommodate strains.  Uploaded values in
spreadsheets are actually raw strains - not strains, which were
causing the form to clear and tests to fail (because of invalid
lookups).  A list of all raw strains is now included in the common
lookup api end point.  The api endpoint test has been updated to
include the new raw_strain element. All previously passing tests
continue to pass.