Added cache to xls_upload formset
The xls_upload formset will require a lot of additional information
to validate many of the dependent fields  - a cache has been created
to hold this information and make it available during cleaning and
validating. The bbox has been moved into the cache, and clean strain
method gets valid values from if depending on selected species.  All
of the previously passing tests continue to pass.