Added cache to xlsx formset and made hatchery option choice field
This commit adds a number of updates to xls tests to include a cache
for spatial fields. Additionally, all of the optional choice fields in
the xlsx_upload have been updated to accommodated null if they are
empty in the spreadsheet, while still validating unknown values. This
commit also includes a new class for cached spatial points to
simplify the api.  All of the previous tests continue to pass,
although the new PointPolygon class has not been tested - either
directly, or when used in the xlsx formset.