Added tests for PointPolygonCollection
This commit includes several tests of the PointsPolygonCollection
class to ensure that it returns the expected results when provided
points fall inside and outside of known spatial entities. All
previously passing tests continue to pass, although there continues to
be several intentionally failing tests that need to be addressed
shortly.