Preparing to push to Beta
This is a big commit that addresses a large number of small issues.
Many of the changes are associated with updates to models and tests
associated with data uploads. All of the js files have built for
production, and the centroids have been updated for all grids, and
management units (no more missing values).  Finally Management Unit has
been added back into the option list (and response) for filtered
stocking events to provide a level of resolution between 10-minute grid
and jurisdiction.  All of previously passing tests continue to pass.