Updated the search by id box in base template
The main menu has a search by id box on left side of the screen. It
did not work for pages that were not lists. This has now been
updated. It defaults to stocking id and redirects to stocking list
unless over ridden.  All previously passing tests continue to pass.