Added age in months to stocking form
This was an oversight that was identified during the stocking
coordinator workshop. The form now includes a field for age in month.
The form has also been updated to accept float values for the average
weight (it was an integer) which often produced errors.  The
associated tests have been updated.  All of the previously passing
tests continue to pass.