Added optional color field to serializer and associated tests
This commit updates the serializers for all of the entities that will
be colours with an optional color attribute. If color=True, the
associated color will be included in the serialized response.  Tests
verify that the optional flag works as expected. This commit also
updates the common and stocking lookup api endpoints to include color
in their response. The associated tests where updated or created. All
previously passing tests continue to pass.