Added button to Find forms to toggle map
The maps widget take up a lot of space on the find event and find cwt
forms. This commit adds a button and some associated JavaScript to the
templates to hide and show the maps. The maps must be shown by default
so that they initialize correctly. All of the previously passing unit
tests continue to pass.